requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
python-telegram-bot==20.7
reportlab==4.0.7
python-dotenv==1.0.0
click==8.1.7
colorama==0.4.6
lxml==4.9.3
webdriver-manager==4.0.1